% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetics.R
\name{build_multi_region_trees}
\alias{build_multi_region_trees}
\title{Build phylogenies from multiple genomic regions}
\usage{
build_multi_region_trees(
  genome_alignment,
  regions,
  method = "K80",
  bootstrap = 100
)
}
\arguments{
\item{genome_alignment}{DNAStringSet of full genomes (aligned)}

\item{regions}{Named list of (start, end) coordinates}

\item{method}{Distance method}

\item{bootstrap}{Number of bootstrap replicates}
}
\value{
Named list of phylo objects
}
\description{
Build phylogenies from multiple genomic regions
}
