% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetics.R
\name{build_phylo_tree}
\alias{build_phylo_tree}
\title{Build phylogenetic tree from DNA sequences}
\usage{
build_phylo_tree(sequences, method = "K80", tree_method = "nj", bootstrap = 0)
}
\arguments{
\item{sequences}{DNAStringSet}

\item{method}{Distance method ("K80", "JC69", "raw", "TN93")}

\item{tree_method}{Tree building ("nj", "upgma")}

\item{bootstrap}{Number of bootstrap replicates (0 = no bootstrap)}
}
\value{
phylo object
}
\description{
Build phylogenetic tree from DNA sequences
}
