% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_analysis.R
\name{calculate_codon_frequencies}
\alias{calculate_codon_frequencies}
\title{Calculate codon frequencies}
\usage{
calculate_codon_frequencies(sequences)
}
\arguments{
\item{sequences}{Character vector or DNAStringSet}
}
\value{
Named vector of codon frequencies
}
\description{
Calculate codon frequencies
}
