% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_analysis.R
\name{calculate_enc}
\alias{calculate_enc}
\title{Calculate Effective Number of Codons (ENC/Nc)
Wright (1990) Gene 87:23-29}
\usage{
calculate_enc(sequence)
}
\arguments{
\item{sequence}{Single sequence (character or DNAString)}
}
\value{
Numeric ENC value (2-61), or NA if sequence too short
}
\description{
Calculate Effective Number of Codons (ENC/Nc)
Wright (1990) Gene 87:23-29
}
