% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimized_analysis.R
\name{calculate_gc_matrix_fast}
\alias{calculate_gc_matrix_fast}
\title{GC content matrix for multiple sequences}
\usage{
calculate_gc_matrix_fast(sequences)
}
\arguments{
\item{sequences}{DNAStringSet or character vector}
}
\value{
Matrix: sequences x c(GC_overall, GC1, GC2, GC3)
}
\description{
GC content matrix for multiple sequences
}
