% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_logos.R
\name{calculate_information_content}
\alias{calculate_information_content}
\title{Information content per position}
\usage{
calculate_information_content(sequences, alphabet = c("A", "C", "G", "T"))
}
\arguments{
\item{sequences}{Character vector of equal-length sequences}

\item{alphabet}{Valid character set}
}
\value{
Numeric vector of information content values
}
\description{
Information content per position
}
