% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimized_analysis.R
\name{calculate_rscu_correct}
\alias{calculate_rscu_correct}
\title{Calculate RSCU using seqinr}
\usage{
calculate_rscu_correct(sequences)
}
\arguments{
\item{sequences}{DNAStringSet or character vector}
}
\value{
Data frame with codon, amino_acid, count, frequency, RSCU
}
\description{
Calculate RSCU using seqinr
}
