% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnazyme_activity.R
\name{CANONICAL_1023_CORE}
\alias{CANONICAL_1023_CORE}
\title{Canonical 10-23 DNAzyme catalytic core (15 nt, 5'->3').}
\usage{
CANONICAL_1023_CORE
}
\description{
Canonical 10-23 DNAzyme catalytic core (15 nt, 5'->3').
}
