% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host_detection.R
\name{classify_host_type}
\alias{classify_host_type}
\title{Classify host as monocot or dicot}
\usage{
classify_host_type(host_name)
}
\arguments{
\item{host_name}{Character string of host name}
}
\value{
"monocot", "dicot", or NA
}
\description{
Classify host as monocot or dicot
}
