% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetics.R
\name{compare_trees}
\alias{compare_trees}
\title{Compare two phylogenetic trees}
\usage{
compare_trees(tree1, tree2, labels = NULL)
}
\arguments{
\item{tree1}{First phylo object}

\item{tree2}{Second phylo object}

\item{labels}{Common tip labels (if NULL, uses intersection)}
}
\value{
List with comparison metrics
}
\description{
Compare two phylogenetic trees
}
