% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_logos.R
\name{create_combined_logo}
\alias{create_combined_logo}
\title{Create combined DNA and protein logo plot}
\usage{
create_combined_logo(dna_sequences, title = "Motif Sequence Logos")
}
\arguments{
\item{dna_sequences}{DNAStringSet}

\item{title}{Main title}
}
\value{
Combined ggplot object
}
\description{
Create combined DNA and protein logo plot
}
