% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_logos.R
\name{create_combined_logos}
\alias{create_combined_logos}
\title{Create combined DNA and protein logos}
\usage{
create_combined_logos(dna_sequences, protein_sequences, title_prefix = "Motif")
}
\arguments{
\item{dna_sequences}{DNAStringSet}

\item{protein_sequences}{AAStringSet or character vector}

\item{title_prefix}{Title prefix string}
}
\value{
Combined ggplot
}
\description{
Create combined DNA and protein logos
}
