% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_logos.R
\name{create_dna_logo}
\alias{create_dna_logo}
\title{Create DNA sequence logo from aligned sequences}
\usage{
create_dna_logo(sequences, title = "DNA Sequence Logo", color_scheme = "auto")
}
\arguments{
\item{sequences}{DNAStringSet or character vector (must be aligned)}

\item{title}{Plot title}

\item{color_scheme}{Color scheme ("auto", "nucleotide", "base_pairing")}
}
\value{
ggplot object
}
\description{
Create DNA sequence logo from aligned sequences
}
