% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host_detection.R
\name{DICOT_CODON_TABLE}
\alias{DICOT_CODON_TABLE}
\title{Dicot codon table (based on Arabidopsis thaliana / Nicotiana tabacum)}
\usage{
DICOT_CODON_TABLE
}
\description{
Dicot codon table (based on Arabidopsis thaliana / Nicotiana tabacum)
}
