% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host_detection.R
\name{.optimal_codons}
\alias{.optimal_codons}
\title{Optimal (most frequent) codon per amino acid for a host table.}
\usage{
.optimal_codons(host_table)
}
\arguments{
\item{host_table}{Named numeric vector of codon frequencies.}
}
\value{
Named character vector mapping amino acid -> optimal codon.
}
\description{
Optimal (most frequent) codon per amino acid for a host table.
}
\keyword{internal}
