% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_analysis.R
\name{extract_codons_from_set}
\alias{extract_codons_from_set}
\title{Extract codons from DNAStringSet}
\usage{
extract_codons_from_set(dna_set)
}
\arguments{
\item{dna_set}{DNAStringSet object}
}
\value{
List of character vectors of codons
}
\description{
Extract codons from DNAStringSet
}
