% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_flanks.R
\name{extract_homologs_with_flanks}
\alias{extract_homologs_with_flanks}
\title{Extract homologous motifs with flanks}
\usage{
extract_homologs_with_flanks(
  sequences,
  homolog_positions,
  flank_5prime = 0,
  flank_3prime = 0
)
}
\arguments{
\item{sequences}{DNAStringSet}

\item{homolog_positions}{Data frame from find_motif_homologs}

\item{flank_5prime}{5' flank length}

\item{flank_3prime}{3' flank length}
}
\value{
DNAStringSet of extracted sequences
}
\description{
Extract homologous motifs with flanks
}
