% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_flanks.R
\name{extract_motif_with_flanks}
\alias{extract_motif_with_flanks}
\title{Extract motif with flanking sequences}
\usage{
extract_motif_with_flanks(
  sequences,
  motif_start,
  motif_length,
  flank_5prime = 0,
  flank_3prime = 0,
  extract_from = "alignment"
)
}
\arguments{
\item{sequences}{DNAStringSet of genomes or aligned sequences}

\item{motif_start}{Start position of motif in first sequence}

\item{motif_length}{Length of core motif}

\item{flank_5prime}{Length of 5' flank}

\item{flank_3prime}{Length of 3' flank}

\item{extract_from}{"alignment" or "unaligned"}
}
\value{
List with motif, 5' flank, 3' flank, and combined sequences
}
\description{
Extract motif with flanking sequences
}
