% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_flanks.R
\name{find_motif_homologs}
\alias{find_motif_homologs}
\title{Find motif homologs in sequences using BLAST-like approach}
\usage{
find_motif_homologs(query_motif, target_sequences, max_mismatches = 2)
}
\arguments{
\item{query_motif}{DNAString of query motif}

\item{target_sequences}{DNAStringSet to search}

\item{max_mismatches}{Maximum allowed mismatches}
}
\value{
Data frame with positions of matches
}
\description{
Find motif homologs in sequences using BLAST-like approach
}
