% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_flanks.R
\name{generate_flank_variants}
\alias{generate_flank_variants}
\title{Generate multiple motif+flank combinations}
\usage{
generate_flank_variants(
  sequences,
  motif_start,
  motif_length,
  flank_sizes = c(0, 10, 20, 30, 50)
)
}
\arguments{
\item{sequences}{DNAStringSet}

\item{motif_start}{Motif start position}

\item{motif_length}{Motif length}

\item{flank_sizes}{Vector of flank sizes to test}
}
\value{
Named list of extraction results
}
\description{
Generate multiple motif+flank combinations
}
