% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_analysis.R
\name{get_all_codons}
\alias{get_all_codons}
\title{Get all codons from multiple sequences}
\usage{
get_all_codons(sequences)
}
\arguments{
\item{sequences}{Character vector or DNAStringSet}
}
\value{
Character vector of all codons
}
\description{
Get all codons from multiple sequences
}
