% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_genbank.R
\name{get_all_proteins}
\alias{get_all_proteins}
\title{Get all protein sequences as a named character vector (FASTA-ready)}
\usage{
get_all_proteins(records)
}
\arguments{
\item{records}{Output of load_genbank_folder()}
}
\value{
Named character vector of protein sequences.
}
\description{
Get all protein sequences as a named character vector (FASTA-ready)
}
