% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_genbank.R
\name{get_all_sequences}
\alias{get_all_sequences}
\title{Get all nucleotide sequences as a named character vector}
\usage{
get_all_sequences(records)
}
\arguments{
\item{records}{Output of load_genbank_folder()}
}
\value{
Named character vector of nucleotide sequences.
}
\description{
Get all nucleotide sequences as a named character vector
}
