% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host_detection.R
\name{get_codon_table_for_host_type}
\alias{get_codon_table_for_host_type}
\title{Get appropriate codon table for a host type}
\usage{
get_codon_table_for_host_type(host_type)
}
\arguments{
\item{host_type}{"monocot", "dicot", or NA}
}
\value{
Named vector of codon frequencies
}
\description{
Get appropriate codon table for a host type
}
