% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_genbank.R
\name{has_genbank_id}
\alias{has_genbank_id}
\title{Does a string contain a GenBank-style accession?}
\usage{
has_genbank_id(x)
}
\arguments{
\item{x}{Character vector to test.}
}
\value{
Logical vector, same length as \code{x}.
}
\description{
Does a string contain a GenBank-style accession?
}
