% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host_detection.R
\name{interactive_host_assignment}
\alias{interactive_host_assignment}
\title{Interactive function to manually assign hosts}
\usage{
interactive_host_assignment(sequence_names, auto_classification)
}
\arguments{
\item{sequence_names}{Character vector of sequence names}

\item{auto_classification}{Data frame from create_host_classification_table}
}
\value{
Updated classification data frame
}
\description{
Interactive function to manually assign hosts
}
