% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host_detection.R
\name{load_host_classification}
\alias{load_host_classification}
\title{Load manually edited host classification}
\usage{
load_host_classification(filepath)
}
\arguments{
\item{filepath}{Path to CSV file}
}
\value{
Data frame with host classification
}
\description{
Load manually edited host classification
}
