% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_logos.R
\name{logo_statistics}
\alias{logo_statistics}
\title{Calculate summary statistics for sequence logo}
\usage{
logo_statistics(sequences)
}
\arguments{
\item{sequences}{DNAStringSet or character vector}
}
\value{
Data frame with position-wise statistics
}
\description{
Calculate summary statistics for sequence logo
}
