% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host_detection.R
\name{MONOCOT_CODON_TABLE}
\alias{MONOCOT_CODON_TABLE}
\title{Monocot codon table (based on Zea mays / Triticum aestivum)}
\usage{
MONOCOT_CODON_TABLE
}
\description{
Monocot codon table (based on Zea mays / Triticum aestivum)
}
