% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_genbank.R
\name{parse_genbank_file}
\alias{parse_genbank_file}
\title{Parse a single GenBank file (may contain multiple // records)}
\usage{
parse_genbank_file(filepath)
}
\arguments{
\item{filepath}{Path to a .gb file.}
}
\value{
List of records.
}
\description{
Parse a single GenBank file (may contain multiple // records)
}
