% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetics.R
\name{phylogenetic_analysis_workflow}
\alias{phylogenetic_analysis_workflow}
\title{Complete phylogenetic analysis workflow}
\usage{
phylogenetic_analysis_workflow(genome_alignment, regions, bootstrap = 100)
}
\arguments{
\item{genome_alignment}{DNAStringSet}

\item{regions}{Named list of genomic regions}

\item{bootstrap}{Number of bootstrap replicates}
}
\value{
List with all trees and comparisons
}
\description{
Complete phylogenetic analysis workflow
}
