% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_cai_by_host}
\alias{plot_cai_by_host}
\title{Violin/box plot of Codon Adaptation Index by host type}
\usage{
plot_cai_by_host(cai_results, host_classification)
}
\arguments{
\item{cai_results}{Data frame from \code{\link[=calculate_host_specific_cai]{calculate_host_specific_cai()}}
(columns \code{isolate}, \code{CAI}).}

\item{host_classification}{Data frame with columns \code{isolate}, \code{host_type}.}
}
\value{
ggplot object.
}
\description{
Violin/box plot of Codon Adaptation Index by host type
}
