% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_logos.R
\name{plot_conservation}
\alias{plot_conservation}
\title{Plot conservation across sequence}
\usage{
plot_conservation(sequences, title = "Sequence Conservation")
}
\arguments{
\item{sequences}{DNAStringSet or character vector}

\item{title}{Plot title}
}
\value{
ggplot object
}
\description{
Plot conservation across sequence
}
