% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_logos.R
\name{plot_information_content}
\alias{plot_information_content}
\title{Plot information content along sequence}
\usage{
plot_information_content(
  sequences,
  title = "Information Content",
  seq_type = "DNA"
)
}
\arguments{
\item{sequences}{Character vector}

\item{title}{Plot title}

\item{seq_type}{"DNA" or "protein"}
}
\value{
ggplot object
}
\description{
Plot information content along sequence
}
