% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_pca_scree}
\alias{plot_pca_scree}
\title{Scree plot of variance explained by each principal component}
\usage{
plot_pca_scree(pca_results)
}
\arguments{
\item{pca_results}{List from \code{\link[=perform_rscu_pca_improved]{perform_rscu_pca_improved()}}.}
}
\value{
ggplot object.
}
\description{
Scree plot of variance explained by each principal component
}
