% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_logos.R
\name{plot_pssm_heatmap}
\alias{plot_pssm_heatmap}
\title{Plot PSSM as heatmap}
\usage{
plot_pssm_heatmap(pssm, title = "Position-Specific Scoring Matrix")
}
\arguments{
\item{pssm}{PSSM matrix (letters x positions)}

\item{title}{Plot title}
}
\value{
ggplot object
}
\description{
Plot PSSM as heatmap
}
