% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetics.R
\name{plot_tanglegram}
\alias{plot_tanglegram}
\title{Create tanglegram comparing two trees}
\usage{
plot_tanglegram(tree1, tree2, main_title = "Tanglegram Comparison")
}
\arguments{
\item{tree1}{First phylo object}

\item{tree2}{Second phylo object}

\item{main_title}{Plot title}
}
\value{
dendextend tanglegram object
}
\description{
Create tanglegram comparing two trees
}
