% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host_detection.R
\name{POTYVIRUS_HOSTS}
\alias{POTYVIRUS_HOSTS}
\title{Common potyvirus hosts with monocot/dicot classification
Based on known potyvirus host ranges}
\usage{
POTYVIRUS_HOSTS
}
\description{
Common potyvirus hosts with monocot/dicot classification
Based on known potyvirus host ranges
}
