% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_logos.R
\name{save_combined_logo}
\alias{save_combined_logo}
\title{Save sequence logo to file}
\usage{
save_combined_logo(
  sequences,
  output_file = "sequence_logo.png",
  width = 12,
  height = 8,
  dpi = 300
)
}
\arguments{
\item{sequences}{DNAStringSet}

\item{output_file}{Output file path}

\item{width}{Plot width}

\item{height}{Plot height}

\item{dpi}{Resolution}
}
\description{
Save sequence logo to file
}
