% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host_detection.R
\name{save_host_classification}
\alias{save_host_classification}
\title{Save host classification table to CSV for manual editing}
\usage{
save_host_classification(host_classification, filepath)
}
\arguments{
\item{host_classification}{Data frame}

\item{filepath}{Output file path}
}
\description{
Save host classification table to CSV for manual editing
}
