% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_flanks.R
\name{save_motif_flanks}
\alias{save_motif_flanks}
\title{Save motif+flank sequences to FASTA files}
\usage{
save_motif_flanks(extraction_result, output_prefix = "motif")
}
\arguments{
\item{extraction_result}{Result from extract_motif_with_flanks}

\item{output_prefix}{Prefix for output files}
}
\description{
Save motif+flank sequences to FASTA files
}
