% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_flanks.R
\name{workflow_motif_extraction}
\alias{workflow_motif_extraction}
\title{Complete motif extraction workflow}
\usage{
workflow_motif_extraction(
  sequences,
  motif_start,
  motif_length,
  flank_variants = c(0, 10, 20, 30),
  find_homologs = TRUE,
  output_dir = "motif_extraction"
)
}
\arguments{
\item{sequences}{DNAStringSet}

\item{motif_start}{Start position}

\item{motif_length}{Length of motif}

\item{flank_variants}{Vector of flank sizes to generate}

\item{find_homologs}{Whether to find homologs}

\item{output_dir}{Output directory}
}
\value{
List of results
}
\description{
Complete motif extraction workflow
}
