% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_genbank.R
\name{write_fasta}
\alias{write_fasta}
\title{Write sequences to FASTA format}
\usage{
write_fasta(seqs, file, width = 70L)
}
\arguments{
\item{seqs}{Named character vector of sequences.}

\item{file}{Output filepath.}

\item{width}{Line wrap width (default 70).}
}
\description{
Write sequences to FASTA format
}
